requirements.txt
Django==2.2.11
mysqlclient==1.4.6
Pillow==7.0.0
django-crispy-forms>=1.6.0
django-import-export>=0.5.1
django-reversion>=2.0.0
django-formtools==2.1
future==0.15.2
httplib2==0.9.2
six==1.10.0
requests
django-simple-captcha
redis
django-pure-pagination
